Move genome search box test to the corresponding file
